Change tile entities to block entities
