+++ src/main/java/com/jumpcutfindo/happyholidays/client/block/MusicBoxRenderer.java
@@ -1,12 +1,12 @@
 package com.jumpcutfindo.happyholidays.client.block;
 
 import com.jumpcutfindo.happyholidays.client.block.model.MusicBoxModel;
-import com.jumpcutfindo.happyholidays.common.tileentity.christmas.MusicBoxTileEntity;
+import com.jumpcutfindo.happyholidays.common.blockentity.christmas.MusicBoxBlockEntity;
 
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import software.bernie.geckolib3.renderers.geo.GeoBlockRenderer;
 
-public class MusicBoxRenderer extends GeoBlockRenderer<MusicBoxTileEntity> {
+public class MusicBoxRenderer extends GeoBlockRenderer<MusicBoxBlockEntity> {
     public MusicBoxRenderer(BlockEntityRendererProvider.Context rendererDispatcherIn) {
         super(rendererDispatcherIn, new MusicBoxModel());
     }
