+++ src/main/java/com/jumpcutfindo/happyholidays/client/block/model/MusicBoxModel.java
@@ -1,24 +1,24 @@
 package com.jumpcutfindo.happyholidays.client.block.model;
 
 import com.jumpcutfindo.happyholidays.HappyHolidaysMod;
-import com.jumpcutfindo.happyholidays.common.tileentity.christmas.MusicBoxTileEntity;
+import com.jumpcutfindo.happyholidays.common.blockentity.christmas.MusicBoxBlockEntity;
 
 import net.minecraft.resources.ResourceLocation;
 import software.bernie.geckolib3.model.AnimatedGeoModel;
 
-public class MusicBoxModel extends AnimatedGeoModel<MusicBoxTileEntity> {
+public class MusicBoxModel extends AnimatedGeoModel<MusicBoxBlockEntity> {
     @Override
-    public ResourceLocation getModelLocation(MusicBoxTileEntity object) {
+    public ResourceLocation getModelLocation(MusicBoxBlockEntity object) {
         return new ResourceLocation(HappyHolidaysMod.MOD_ID, "geo/christmas/music_box.geo.json");
     }
 
     @Override
-    public ResourceLocation getTextureLocation(MusicBoxTileEntity object) {
+    public ResourceLocation getTextureLocation(MusicBoxBlockEntity object) {
         return new ResourceLocation(HappyHolidaysMod.MOD_ID, "textures/block/music_box.png");
     }
 
     @Override
-    public ResourceLocation getAnimationFileLocation(MusicBoxTileEntity animatable) {
+    public ResourceLocation getAnimationFileLocation(MusicBoxBlockEntity animatable) {
         return new ResourceLocation(HappyHolidaysMod.MOD_ID, "animations/christmas/music_box.animation.json");
     }
 }
