+++ src/main/java/com/jumpcutfindo/happyholidays/common/blockentity/HappyHolidaysEntityBlock.java
@@ -1,0 +1,4 @@
+package com.jumpcutfindo.happyholidays.common.blockentity;
+
+public interface HappyHolidaysEntityBlock {
+}
