+++ src/main/java/com/jumpcutfindo/happyholidays/common/blockentity/christmas/ChristmasEntityBlock.java
@@ -1,0 +1,6 @@
+package com.jumpcutfindo.happyholidays.common.blockentity.christmas;
+
+import com.jumpcutfindo.happyholidays.common.blockentity.HappyHolidaysEntityBlock;
+
+public interface ChristmasEntityBlock extends HappyHolidaysEntityBlock {
+}
