+++ src/main/java/com/jumpcutfindo/happyholidays/common/container/christmas/star/ChristmasStarSlot.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import com.jumpcutfindo.happyholidays.common.events.christmas.ChristmasStarEvent;
 import com.jumpcutfindo.happyholidays.common.registry.christmas.ChristmasItems;
-import com.jumpcutfindo.happyholidays.common.tileentity.christmas.ChristmasStarTileEntity;
+import com.jumpcutfindo.happyholidays.common.blockentity.christmas.ChristmasStarBlockEntity;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
@@ -16,11 +16,11 @@ import net.minecraft.world.level.Level;
 import net.minecraftforge.common.MinecraftForge;
 
 public class ChristmasStarSlot extends Slot {
-    private final ChristmasStarTileEntity starTileEntity;
+    private final ChristmasStarBlockEntity starBlockEntity;
 
-    public ChristmasStarSlot(ChristmasStarTileEntity starTileEntity, int p_i1824_2_, int p_i1824_3_, int p_i1824_4_) {
-        super(starTileEntity, p_i1824_2_, p_i1824_3_, p_i1824_4_);
-        this.starTileEntity = starTileEntity;
+    public ChristmasStarSlot(ChristmasStarBlockEntity starBlockEntity, int p_i1824_2_, int p_i1824_3_, int p_i1824_4_) {
+        super(starBlockEntity, p_i1824_2_, p_i1824_3_, p_i1824_4_);
+        this.starBlockEntity = starBlockEntity;
     }
 
     @Override
@@ -28,15 +28,15 @@ public class ChristmasStarSlot extends Slot {
         boolean flag = ChristmasItems.isOrnamentItem(itemStack);
 
         if (flag) {
-            Level level = starTileEntity.getLevel();
+            Level level = starBlockEntity.getLevel();
             if (level != null && !level.isClientSide()) {
-                BlockPos blockPos = starTileEntity.getBlockPos();
+                BlockPos blockPos = starBlockEntity.getBlockPos();
 
                 Player playerEntity = level.getNearestPlayer(blockPos.getX(), blockPos.getY(), blockPos.getZ(),
                         10.0D, false);
 
                 if (playerEntity != null) {
-                    ChristmasStarEvent event = new ChristmasStarEvent.PutOrnament(starTileEntity, playerEntity);
+                    ChristmasStarEvent event = new ChristmasStarEvent.PutOrnament(starBlockEntity, playerEntity);
                     MinecraftForge.EVENT_BUS.post(event);
                 }
             }
@@ -47,9 +47,9 @@ public class ChristmasStarSlot extends Slot {
 
     @Override
     public boolean mayPickup(Player playerEntity) {
-        boolean mayPickup = !this.starTileEntity.isBonusActive();
+        boolean mayPickup = !this.starBlockEntity.isBonusActive();
 
-        if (!mayPickup && starTileEntity.getLevel() != null && starTileEntity.getLevel().isClientSide())
+        if (!mayPickup && starBlockEntity.getLevel() != null && starBlockEntity.getLevel().isClientSide())
             playerEntity.sendMessage(
                 new TranslatableComponent("chat.happyholidays.christmas_star.bonus_still_active")
                         .withStyle(ChatFormatting.RED),
