+++ src/main/java/com/jumpcutfindo/happyholidays/common/events/christmas/ChristmasStarEvent.java
@@ -1,42 +1,42 @@
 package com.jumpcutfindo.happyholidays.common.events.christmas;
 
-import com.jumpcutfindo.happyholidays.common.tileentity.christmas.ChristmasStarTileEntity;
+import com.jumpcutfindo.happyholidays.common.blockentity.christmas.ChristmasStarBlockEntity;
 
 import net.minecraft.world.entity.player.Player;
 
 public class ChristmasStarEvent extends ChristmasEvent {
-    final ChristmasStarTileEntity tileEntity;
+    final ChristmasStarBlockEntity blockEntity;
 
-    public ChristmasStarEvent(ChristmasStarTileEntity tileEntity, Player playerEntity) {
+    public ChristmasStarEvent(ChristmasStarBlockEntity blockEntity, Player playerEntity) {
         super(playerEntity);
-        this.tileEntity = tileEntity;
+        this.blockEntity = blockEntity;
     }
 
     public static class PutOrnament extends ChristmasStarEvent {
-        public PutOrnament(ChristmasStarTileEntity tileEntity, Player playerEntity) {
-            super(tileEntity, playerEntity);
+        public PutOrnament(ChristmasStarBlockEntity blockEntity, Player playerEntity) {
+            super(blockEntity, playerEntity);
         }
     }
 
     public static class IncreaseTier extends ChristmasStarEvent {
-        public IncreaseTier(ChristmasStarTileEntity tileEntity, Player playerEntity) {
-            super(tileEntity, playerEntity);
+        public IncreaseTier(ChristmasStarBlockEntity blockEntity, Player playerEntity) {
+            super(blockEntity, playerEntity);
         }
 
         public int getTier() {
-            return tileEntity.getCurrentTier();
+            return blockEntity.getCurrentTier();
         }
     }
 
     public static class SummonSanta extends ChristmasStarEvent {
-        public SummonSanta(ChristmasStarTileEntity tileEntity, Player playerEntity) {
-            super(tileEntity, playerEntity);
+        public SummonSanta(ChristmasStarBlockEntity blockEntity, Player playerEntity) {
+            super(blockEntity, playerEntity);
         }
     }
 
     public static class ReachBonus extends ChristmasStarEvent {
-        public ReachBonus(ChristmasStarTileEntity tileEntity, Player playerEntity) {
-            super(tileEntity, playerEntity);
+        public ReachBonus(ChristmasStarBlockEntity blockEntity, Player playerEntity) {
+            super(blockEntity, playerEntity);
         }
     }
 }
