--- src/main/java/com/jumpcutfindo/happyholidays/common/tileentity/HappyHolidaysTileEntity.java
@@ -1,4 +1,0 @@
-package com.jumpcutfindo.happyholidays.common.tileentity;
-
-public interface HappyHolidaysTileEntity {
-}
