--- src/main/java/com/jumpcutfindo/happyholidays/common/tileentity/christmas/IChristmasTileEntity.java
@@ -1,6 +1,0 @@
-package com.jumpcutfindo.happyholidays.common.tileentity.christmas;
-
-import com.jumpcutfindo.happyholidays.common.tileentity.HappyHolidaysTileEntity;
-
-public interface IChristmasTileEntity extends HappyHolidaysTileEntity {
-}
